Switch screenshot trigger implementation
Originally screenshots were triggered by enabling a feature on the
diagnostic menu that changed the behavior of the blackout switch. This
was convenient, but a bit awkward, and probably not the greatest idea
when an alternative exists.

The new trigger is to simply support the "Ctrl-PrintScreen" key sequence
on an attached USB keyboard, without needing any special modes to be
enabled.