Fix USB MSC device initialization
This is to deal with a change in the upstream CherryUSB library, where
some initialization steps now require an explicit function call. Without
this change, MSC wasn't working anymore.

The intent of this change is to make it possible to move MSC device
setup out of the enumeration task.  Right now we're circumventing the
intent of that by not having a separate task for MSC devices. It may
make sense to remedy this in the future.