Remove meter probe config import/export code
There is no valid use case, outside of a development environment, to
want to import/export this data from the enlarging timer. Its more
likely to end up letting a user corrupt their hard-to-fix probe
calibrations than provide any real benefit.

From here on out, most peripheral calibrations should be considered as
factory functions performed with external software. Some of this
software may be published in the future.

The only exception is DensiStick target calibration, which should be
do-able on device using reference materials likely to be provided to
users.

Ref #82