Change piezo buzzer component
This changes the piezo buzzer on the main board from a Murata
PKM22EPPH4001-B0 to a Mallory PT-2040PQ.

Both these components have the identical footprint, and we don't have a
good model for the Mallory part, so the only change happening here is
BOM metadata.

This change is due to Murata discontinuing their entire piezo product
line.  The new component is electrically and mechanically a drop-in
replacement, but its frequency/volume behavior does differ somewhat.

Fix #86