Change DMX connector part from NC5FBH to NC5FAH
This is a drop-in replacement, and all of the routing is unchanged.
However, KiCad's library does make a few slight silkscreen tweaks for
the other part.  Electrically and mechanically, everything is unchanged.

Still using the rather rough looking CAD model for the old part, but
that may be changed when/if Neutrik releases a better model.  (They are
currently in the middle of a product line refresh, and have already
released better models for the other parts.)

The motivator for this change is partly cost, and partly the discovery
that no DMX512 cable seems to bind its connector shells to anything. As
such, having a 360 degree shield ground connection serves no useful
purpose.

Fix #84